refactor(curation): extract shared model-response JSON parser
`curation.evaluate._extract_json` and
`organizations.services.ai_descriptions._parse_response` duplicated
the same "strip markdown fences, try json.loads, fall back to brace
extraction" logic. Extract the shared half into
`curation.json_utils.extract_json` so fixes land in one place.

Side benefit: ai_descriptions inherits the depth-aware brace
matcher (`_find_json_end`), which correctly handles trailing prose
containing stray `}` — the old `text.rfind("}")` fallback would
misbehave there.

`_parse_response` shrinks to its actual job: call the shared
extractor, validate description/action_text fields, construct
DraftedCopy. Moves the `TestExtractJson` tests to
`test_json_utils.py` and adds a regression test for the stray-brace
case.
